refactor and organize directory
