+++ cmd/main/indexer.go
@@ -1,214 +1,30 @@
 package main
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"log"
 	"net/http"
 	_ "net/http/pprof"
 	"os"
-	"path/filepath"
-	"strings"
 	"sync"
 	"time"
+
+	flFuntion "indexer_project/cmd/main/pkg/fileProcessor"
+	zsAPI "indexer_project/cmd/main/pkg/zincSearchApi"
 )
 
-// type mail struct {
-// 	Message_ID                string
-// 	Date                      string
-// 	From                      string
-// 	to                        string
-// 	subject                   string
-// 	Mime_Version              string
-// 	Content_Type              string
-// 	Content_Transfer_Encoding string
-// 	X_From                    string
-// 	X_To                      string
-// 	X_cc                      string
-// 	X_bcc                     string
-// 	X_Folder                  string
-// 	X_Origin                  string
-// 	X_FileName                string
-// 	content                   string
-// }
-
-type Mailstruct struct {
-	fieldA string
-	fieldB []map[string]any
-}
-
-type MailJson struct {
-	FieldA string           `json:"index"`
-	FieldB []map[string]any `json:"records"`
-}
-
-func (t *Mailstruct) MarshalJSON() ([]byte, error) {
-	return json.Marshal(MailJson{
-		t.fieldA,
-		t.fieldB,
-	})
-}
-
-func zinSearch_upload(data []byte) {
-
-	user := "admin"
-	password := "Complexpass#123"
-
-	// index := "maildir"
-	zinc_host := "http://localhost:4080"
-	zinc_url := zinc_host + "/api/_bulkv2"
-	counter := 0
-
-	req, err := http.NewRequest("POST", zinc_url, bytes.NewBuffer(data))
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	req.SetBasicAuth(user, password)
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("User-Agent", "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_4) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/81.0.4044.138 Safari/537.36")
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode == 200 {
-		counter += 1
-	} else {
-		log.Println(resp.StatusCode)
-	}
-
-	fmt.Println("ok status: ", counter)
-
-}
-
-func get_mail_map_format(fileContent string) map[string]any {
-
-	part_of_mail_list := []string{"Message-ID", "Date", "From", "To", "Subject", "Mime-Version", "Content-Type",
-		"Content-Transfer-Encoding", "X-From", "X-To", "X-cc", "X-bcc", "X-Folder", "X-Origin", "X-FileName", "body"}
-
-	map_mail := make(map[string]any)
-
-	pos_final := 0
-	pos_ini := 0
-
-	for index_part_of_mail, part_of_mail := range part_of_mail_list {
-
-		part_of_mail_2 := part_of_mail + ":"
-
-		if index_part_of_mail < 14 {
-
-			pos_ini = strings.Index(fileContent, part_of_mail_2) + len(part_of_mail_2)
-			pos_final = strings.Index(fileContent, part_of_mail_list[index_part_of_mail+1])
-
-		} else {
-			if index_part_of_mail == 14 {
-				pos_ini = strings.Index(fileContent, part_of_mail_2) + len(part_of_mail_2)
-				pos_final = strings.Index(fileContent, ".nsf")
-				if pos_final == -1 {
-					pos_final = strings.Index(fileContent, ".pst")
-				}
-				pos_final += 4
-			}
-
-			if index_part_of_mail == 15 {
-				pos_ini = strings.Index(fileContent, ".pst")
-				if pos_ini == -1 {
-					pos_ini = strings.Index(fileContent, ".nsf")
-				}
-				pos_ini += 4
-
-			}
-		}
-
-		if pos_final >= pos_ini {
-			if index_part_of_mail != 15 {
-				map_mail[part_of_mail] = strings.TrimSpace(fileContent[pos_ini:pos_final])
-			} else {
-				map_mail[part_of_mail] = strings.TrimSpace(fileContent[pos_ini:])
-			}
-		} else {
-			map_mail[part_of_mail] = ""
-		}
-	}
-
-	return map_mail
-
-}
-
-func get_mails(pathMailDir string) []string {
-	files := []string{}
-
-	//walk throught all directory ignoring all diferent from a file
-	err := filepath.Walk(pathMailDir, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			fmt.Println(err)
-			return err
-		}
-
-		if !info.IsDir() {
-			files = append(files, path)
-		}
-
-		return nil
-	})
-
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	return files
-}
-
-func files_to_json_format(listFiles []string) []byte {
-
-	list_of_mail := []map[string]any{}
-
-	for _, j := range listFiles {
-
-		input_file := j
-		// read the whole content of file and pass it to file variable, in case of error pass it to err variable
-		file, err := ioutil.ReadFile(input_file)
-
-		if err != nil {
-			fmt.Printf("Could not read the file due to this %s error \n", err)
-		} else {
-
-			// if i == 100000 {
-			// 	break
-			// }
-
-			// convert the file into a map format and put into a list of map
-			file_content := string(file)
-			mail_map_format := get_mail_map_format(file_content)
-			list_of_mail = append(list_of_mail, mail_map_format)
-		}
-
-	}
-
-	//convert the structure in a Json format
-	json_file, err := json.Marshal(&Mailstruct{"maildir", list_of_mail})
-	// _ = ioutil.WriteFile("test.json", json_file, 0644)
-
-	if err != nil {
-		fmt.Printf("Error: %s", err.Error())
-	}
-	return json_file
-}
-
-func run(list_of_files []string, wg *sync.WaitGroup) {
+// excute the process to bulk the directory
+// first convert the files into a json structur
+// then call api zincsearch to upload the data
+func run(index string, list_of_files []string, wg *sync.WaitGroup) {
 	defer wg.Done()
 
 	ji := time.Now()
-	json_mails := files_to_json_format(list_of_files)
+	//convert every file into a json format and put into a list
+	json_mails_list := flFuntion.Files_to_json_format(index, list_of_files)
 
-	if len(json_mails) > 0 {
-		zinSearch_upload(json_mails)
+	if len(json_mails_list) > 0 {
+		//call zinSearch API to make data bulk
+		zsAPI.ZinSearch_upload(json_mails_list)
 	}
 
 	ft := time.Since(ji)
@@ -219,6 +35,14 @@ func run(list_of_files []string, wg *sync.WaitGroup) {
 }
 
 func main() {
+	pathMailDir := ""
+	index := "maildir"
+
+	if len(os.Args) > 1 {
+		fmt.Println(os.Args[1])
+		pathMailDir = "./" + os.Args[1] + "/" + index + "/allen-p/sent_mail/"
+
+	}
 
 	var wg sync.WaitGroup
 
@@ -229,33 +53,23 @@ func main() {
 	wg.Add(1) // pprof - so we won't exit prematurely
 
 	fmt.Printf("start: %v\n", time.Now())
-
-	pathMailDir := "./enron_mail_20110402/maildir/"
-
 	mi := time.Now()
-	list_of_files := get_mails(pathMailDir)
+	list_of_files := flFuntion.Get_all_files(pathMailDir)
 	mt := time.Since(mi)
+
 	// printing the time in string format
-	fmt.Println("execution get mail directory time: ", mt.String())
+	fmt.Println("execution get_all_files time: ", mt.String())
 
-	// tamanio = len(list_of_files)
+	limit := len(list_of_files) / 2
 
-	list_of_files_1 := list_of_files[0:100000]
-	list_of_files_2 := list_of_files[100000:200000]
-	list_of_files_3 := list_of_files[200000:300000]
-	list_of_files_4 := list_of_files[300000:400000]
-	list_of_files_5 := list_of_files[400000:]
+	list_of_files_1 := list_of_files[0:limit]
+	list_of_files_2 := list_of_files[limit:]
 
-	wg.Add(1) // for mail process
-	go run(list_of_files_1, &wg)
-	wg.Add(1) // for mail process
-	go run(list_of_files_2, &wg)
-	wg.Add(1) // for mail process
-	go run(list_of_files_3, &wg)
-	wg.Add(1) // for mail process
-	go run(list_of_files_4, &wg)
-	wg.Add(1) // for mail process
-	go run(list_of_files_5, &wg)
+	wg.Add(1)
+	go run(index, list_of_files_1, &wg)
+
+	wg.Add(1)
+	go run(index, list_of_files_2, &wg)
 
 	wg.Wait()
