+++ cmd/main/pkg/envVariables/variables.env
@@ -1,0 +1,3 @@
+USER_ZINC = admin
+PASS_ZINC = Complexpass#123
+HOST_ZINC = http://localhost:4080
