+++ go.mod
@@ -6,5 +6,6 @@ require (
 	github.com/chzyer/readline v1.5.1 // indirect
 	github.com/google/pprof v0.0.0-20221219190121-3cb0bae90811 // indirect
 	github.com/ianlancetaylor/demangle v0.0.0-20220517205856-0058ec4f073c // indirect
+	github.com/joho/godotenv v1.4.0
 	golang.org/x/sys v0.3.0 // indirect
 )
